# Accelerate water dynamics by removing WaterTable resource