--- a/emergence_lib/src/graphics/terrain.rs
+++ b/emergence_lib/src/graphics/terrain.rs
@@ -2,9 +2,9 @@
 
 use crate::{
     items::inventory::InventoryState,
-    simulation::geometry::{MapGeometry, TilePos},
+    simulation::geometry::{Height, MapGeometry, TilePos},
     terrain::{litter::Litter, terrain_assets::TerrainHandles},
-    water::WaterTable,
+    water::WaterDepth,
 };
 use bevy::{prelude::*, utils::HashMap};
 
@@ -15,10 +15,10 @@ pub(super) fn render_litter_piles(
     mut current_ground_litter_piles: Local<HashMap<TilePos, (InventoryState, Entity)>>,
     mut current_floating_litter_piles: Local<HashMap<TilePos, (InventoryState, Entity)>>,
     terrain_query: Query<(Entity, &TilePos, Ref<Litter>)>,
+    water_height_query: Query<(&WaterDepth, &Height)>,
     // PERF: we could add a marker component to improve parallelism
     mut floating_litter_query: Query<&mut Transform>,
     mut commands: Commands,
-    water_table: Res<WaterTable>,
     map_geometry: Res<MapGeometry>,
 ) {
     for (terrain_entity, &tile_pos, litter) in terrain_query.iter() {
@@ -81,8 +81,12 @@ pub(super) fn render_litter_piles(
                     scene: scene_handle.clone(),
                     // This can't be a child of the terrain entity because it needs to be able to
                     // change heights with the water.
-                    transform: floating_litter_transform(tile_pos, &water_table, &map_geometry)
-                        .unwrap_or_default(),
+                    transform: floating_litter_transform(
+                        tile_pos,
+                        &water_height_query,
+                        &map_geometry,
+                    )
+                    .unwrap_or_default(),
                     ..Default::default()
                 })
                 .id();
@@ -95,7 +99,7 @@ pub(super) fn render_litter_piles(
     for (tile_pos, (_, entity)) in current_floating_litter_piles.iter() {
         if let Ok(mut transform) = floating_litter_query.get_mut(*entity) {
             if let Ok(new_transform) =
-                floating_litter_transform(*tile_pos, &water_table, &map_geometry)
+                floating_litter_transform(*tile_pos, &water_height_query, &map_geometry)
             {
                 *transform = new_transform;
             } else {
@@ -108,13 +112,15 @@ pub(super) fn render_litter_piles(
 /// Computes the [`Transform`] for a floating litter entity.
 fn floating_litter_transform(
     tile_pos: TilePos,
-    water_table: &WaterTable,
+    water_height_query: &Query<(&WaterDepth, &Height)>,
     map_geometry: &MapGeometry,
 ) -> Result<Transform, ()> {
     let mut transform = Transform::from_translation(tile_pos.into_world_pos(map_geometry));
-    let terrain_height = map_geometry.get_height(tile_pos).unwrap();
-    let water_depth = water_table.surface_water_depth(tile_pos);
-    let desired_height = terrain_height + water_depth;
+    let Some(terrain_entity) = map_geometry.get_terrain(tile_pos) else {
+        return Err(());
+    };
+    let (water_depth, &terrain_height) = water_height_query.get(terrain_entity).unwrap();
+    let desired_height = water_depth.surface_height(terrain_height);
 
     transform.translation.y = desired_height.into_world_pos();
     Ok(transform)