Fixes #925.

> compute_water_depth     time:   [52.697 µs 53.378 µs 54.102 µs]
>                         change: [-94.370% -94.296% -94.231%] (p = 0.00 < 0.05)
>                         Performance has improved.
> Found 2 outliers among 100 measurements (2.00%)
>   2 (2.00%) low mild
> 
> lateral_water_movement  time:   [12.047 ms 12.084 ms 12.140 ms]
>                         change: [-11.951% -11.285% -10.691%] (p = 0.00 < 0.05)
>                         Performance has improved.
> Found 10 outliers among 100 measurements (10.00%)
>   1 (1.00%) low mild
>   2 (2.00%) high mild
>   7 (7.00%) high severe

# Status

- [x] remove SoilWaterFlowRate lookup (14% improvement on lateral water benchmark)
- [x] remove height lookup and soil water lookup in `update_water_height` (74% improvement on water depth benchmark)
- [x] finish removing the WaterTable resource (12% improvement on lateral water movement benchmark)